wrpll: convert tests to self-checking unittests
