+++ artiq/gateware/drtio/wrpll/ddmtd.py
@@ -219,78 +219,3 @@ class Collector(Module):
             NextValue(self.out_stb, 1),
             NextState("IDLE")
         )
-
-
-def test_collector_main():
-    N = 2
-    collector = Collector(N=N)
-    # check collector phase unwrapping
-    tags = [(0, 0, 0),
-            (0, 1, 1),
-            (2, 1, -1),
-            (3, 1, -2),
-            (0, 1, -3),
-            (1, 1, -4),
-            (2, 1, -5),
-            (3, 1, -6),
-            (3, 3, -4),
-            (0, 0, -4),
-            (0, 1, -3),
-            (0, 2, -2),
-            (0, 3, -1),
-            (0, 0, 0)]
-    for i in range(10):
-        tags.append((i % (2**N), (i+1) % (2**N), 1))
-
-    def generator():
-        for tag_ref, tag_main, out in tags:
-            yield collector.tag_ref.eq(tag_ref)
-            yield collector.tag_main.eq(tag_main)
-            yield collector.main_stb.eq(1)
-            yield collector.ref_stb.eq(1)
-
-            yield
-
-            yield collector.main_stb.eq(0)
-            yield collector.ref_stb.eq(0)
-
-            while not (yield collector.out_stb):
-                yield
-
-            out_main = yield collector.out_main
-            assert out_main == out
-
-    run_simulation(collector, generator())
-
-
-def test_collector_helper():
-    N = 3
-    collector = Collector(N=N)
-    # check collector phase unwrapping
-    tags = [((2**N - 1 - tag) % (2**N), 0) for tag in range(20)]
-    tags += [((tags[-1][0] + 1 + tag) % (2**N), 2) for tag in range(20)]
-    tags += [((tags[-1][0] - 2 - 2*tag) % (2**N), -1) for tag in range(20)]
-
-    def generator():
-        for tag_ref, out in tags:
-            yield collector.tag_ref.eq(tag_ref)
-            yield collector.main_stb.eq(1)
-            yield collector.ref_stb.eq(1)
-
-            yield
-
-            yield collector.main_stb.eq(0)
-            yield collector.ref_stb.eq(0)
-
-            while not (yield collector.out_stb):
-                yield
-
-            out_helper = yield collector.out_helper
-            assert out_helper == out
-
-    run_simulation(collector, generator())
-
-
-if __name__ == "__main__":
-    test_collector_main()
-    test_collector_helper()
