+++ artiq/gateware/drtio/wrpll/thls.py
@@ -612,50 +612,3 @@ def make(function, **kwargs):
     cp = compile(proc, function)
     cp.dimension_processor()
     return proc.implement(cp.encode(), cp.data)
-
-
-a = 0
-b = 0
-c = 0
-
-def foo(x):
-    global a, b, c
-    c = b
-    b = a
-    a = x
-    return 4748*a + 259*b - 155*c
-
-
-def simple_test(x):
-    global a
-    a = a + (x*4 >> 1)
-    return a
-
-
-if __name__ == "__main__":
-    proc = Processor()
-    cp = compile(proc, simple_test)
-    cp.pretty_print()
-    cp.dimension_processor()
-    print(cp.encode())
-    proc_impl = proc.implement(cp.encode(), cp.data)
-
-    def send_values(values):
-        for value in values:
-            yield proc_impl.input.eq(value)
-            yield proc_impl.input_stb.eq(1)
-            yield
-            yield proc_impl.input.eq(0)
-            yield proc_impl.input_stb.eq(0)
-            yield
-            while (yield proc_impl.busy):
-                yield
-    @passive
-    def receive_values(callback):
-        while True:
-            while not (yield proc_impl.output_stb):
-                yield
-            callback((yield proc_impl.output))
-            yield
-
-    run_simulation(proc_impl, [send_values([42, 40, 10, 10]), receive_values(print)])
